Updated README.md
Updated the project overview. The changes here were to explain the purpose of the dog_app.ipynb notebook. 
Updated the project instructions. The changes here were to redirect to my repo and some minor changes on how to setup the environment for further changes.
Added a Walkthrough the dog_app.ipynb notebook. This will be used to explain to the reader what each section of the notebook is.